pyproject.toml
[tool.poetry]
name = "telegram-rag-chatbot"
version = "0.1.0"
description = ""
authors = ["Zarif Aziz <zarifaziz08@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.4"
pydantic = "^2.8.2"
openai = "^1.37.1"
requests = "^2.32.3"
pydantic-settings = "^2.4.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
